+++ extensions/nestedadapter/src/main/kotlin/smartadapter/nestedadapter/SmartNestedRecyclerViewHolder.kt
@@ -1,0 +1,8 @@
+package smartadapter.nestedadapter
+
+import androidx.recyclerview.widget.RecyclerView
+
+interface SmartNestedRecyclerViewHolder {
+
+    val recyclerView: RecyclerView
+}