+++ extensions/viewevent/CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.0.0-beta01
+
+Bumped to beta
+
+### Fixed
+* Fix dragFlags for DragAndDropEventBinder depending on recyclerView is LinearLayoutManager
+* Fix potential crash on collections item position swap
+
 # v1.0.0-alpha02
 
 ### Fixed